requirements.txt
langgraph>=0.2
langchain>=0.3
langchain-core>=0.3
langchain-ollama>=0.1
langsmith>=0.1
python-dotenv
